Rename framebuffer module to render_pass, move/rename other things
