+++ examples/src/bin/deferred/frame/system.rs
@@ -16,10 +16,10 @@ use vulkano::command_buffer::CommandBuffer;
 use vulkano::command_buffer::SubpassContents;
 use vulkano::device::Queue;
 use vulkano::format::Format;
-use vulkano::framebuffer::Framebuffer;
-use vulkano::framebuffer::FramebufferAbstract;
-use vulkano::framebuffer::RenderPass;
-use vulkano::framebuffer::Subpass;
+use vulkano::render_pass::Framebuffer;
+use vulkano::render_pass::FramebufferAbstract;
+use vulkano::render_pass::RenderPass;
+use vulkano::render_pass::Subpass;
 use vulkano::image::view::ImageView;
 use vulkano::image::AttachmentImage;
 use vulkano::image::ImageUsage;
