+++ examples/src/bin/deferred/triangle_draw_system.rs
@@ -13,7 +13,7 @@ use vulkano::command_buffer::AutoCommandBuffer;
 use vulkano::command_buffer::AutoCommandBufferBuilder;
 use vulkano::command_buffer::DynamicState;
 use vulkano::device::Queue;
-use vulkano::framebuffer::Subpass;
+use vulkano::render_pass::Subpass;
 use vulkano::pipeline::viewport::Viewport;
 use vulkano::pipeline::GraphicsPipeline;
 use vulkano::pipeline::GraphicsPipelineAbstract;
