+++ vulkano/src/command_buffer/auto.rs
@@ -46,11 +46,11 @@ use crate::format::AcceptsPixels;
 use crate::format::ClearValue;
 use crate::format::Format;
 use crate::format::FormatTy;
-use crate::framebuffer::Framebuffer;
-use crate::framebuffer::FramebufferAbstract;
-use crate::framebuffer::LoadOp;
-use crate::framebuffer::RenderPass;
-use crate::framebuffer::Subpass;
+use crate::render_pass::Framebuffer;
+use crate::render_pass::FramebufferAbstract;
+use crate::render_pass::LoadOp;
+use crate::render_pass::RenderPass;
+use crate::render_pass::Subpass;
 use crate::image::ImageAccess;
 use crate::image::ImageLayout;
 use crate::instance::QueueFamily;
