+++ vulkano/src/command_buffer/mod.rs
@@ -97,7 +97,7 @@ pub use self::state_cacher::StateCacherOutcome;
 pub use self::traits::CommandBuffer;
 pub use self::traits::CommandBufferExecError;
 pub use self::traits::CommandBufferExecFuture;
-use crate::framebuffer::{Framebuffer, Subpass};
+use crate::render_pass::{Framebuffer, Subpass};
 use crate::pipeline::depth_stencil::DynamicStencilValue;
 use crate::pipeline::viewport::{Scissor, Viewport};
 use crate::query::QueryPipelineStatisticFlags;
