+++ vulkano/src/pipeline/graphics_pipeline/mod.rs
@@ -18,8 +18,8 @@ use crate::descriptor::pipeline_layout::PipelineLayoutSys;
 use crate::descriptor::PipelineLayoutAbstract;
 use crate::device::Device;
 use crate::device::DeviceOwned;
-use crate::framebuffer::RenderPass;
-use crate::framebuffer::Subpass;
+use crate::render_pass::RenderPass;
+use crate::render_pass::Subpass;
 use crate::pipeline::shader::EmptyEntryPointDummy;
 use crate::pipeline::vertex::BufferlessDefinition;
 use crate::pipeline::vertex::IncompatibleVertexDefinitionError;
