+++ vulkano/src/render_pass/compat_atch.rs
@@ -11,7 +11,7 @@
 //! an image view can be used as a render pass attachment.
 
 use crate::format::Format;
-use crate::framebuffer::RenderPassDesc;
+use crate::render_pass::RenderPassDesc;
 use crate::image::view::ImageViewAbstract;
 use std::error;
 use std::fmt;
@@ -178,7 +178,7 @@ mod tests {
     use super::ensure_image_view_compatible;
     use super::IncompatibleRenderPassAttachmentError;
     use crate::format::Format;
-    use crate::framebuffer::RenderPassDesc;
+    use crate::render_pass::RenderPassDesc;
     use crate::image::view::ImageView;
     use crate::image::AttachmentImage;
